+++ test/fixtures/cjs.cjs
@@ -1,5 +1,5 @@
 const yolo = require('fs');
-
+// eslint-disable-next-line no-console
 module.exports = {
   presets: [[ '@babel/preset-env', {
     bugfixes: true,